feature(userInfo) added forms, buttons, and static info
