+++ client/components/index.js
@@ -13,6 +13,7 @@ export {default as CheckOut} from './checkOut'
 
 export {default as UserHome} from './user/user-home'
 export {default as UserInfo} from './user/user-info'
+export {default as UserEdit} from './user/user-edit'
 export {default as UserOrder} from './user/user-order'
 
 // Single home components
