+++ client/components/user/user-edit.js
@@ -1,0 +1,203 @@
+import React from 'react'
+import {connect} from 'react-redux'
+
+const UserInfo = props => {
+  const {
+    firstName,
+    lastName,
+    email,
+    address,
+    address2,
+    city,
+    stateAdd,
+    zip
+  } = props
+
+  const handleChange = event => {
+    console.log(event.target.value)
+  }
+
+  const handleSubmit = event => {
+    console.log(event.target.value)
+  }
+
+  return (
+    <div className="row userInfo" onChange={handleChange}>
+      <form className="col s12">
+        <div className="row">
+          <div className="input-field col s6">
+            <label htmlFor="firstName" className="active">
+              First Name
+            </label>
+            <input
+              onChange={handleChange}
+              value={firstName}
+              id="first_name"
+              type="text"
+              className="validate"
+            />
+          </div>
+          <div className="input-field col s6">
+            <label htmlFor="lastName" className="active">
+              Last Name
+            </label>
+            <input
+              onChange={handleChange}
+              value={lastName}
+              id="last_name"
+              type="text"
+              className="validate"
+            />
+          </div>
+        </div>
+        <div className="row">
+          <div className="input-field col s6">
+            <label htmlFor="email" className="active">
+              Email
+            </label>
+            <input
+              onChange={handleChange}
+              value={email}
+              id="email"
+              type="text"
+              className="validate"
+            />
+          </div>
+        </div>
+        <div className="row">
+          <div className="input-field col s6">
+            <label htmlFor="address1" className="active">
+              Address 1
+            </label>
+            <input
+              onChange={handleChange}
+              value={address}
+              id="address1"
+              type="text"
+              className="validate"
+            />
+          </div>
+          <div className="input-field col s6">
+            <label htmlFor="address2" className="active">
+              Address 2
+            </label>
+            <input
+              onChange={handleChange}
+              value={address2}
+              id="address2"
+              type="text"
+              className="validate"
+            />
+          </div>
+        </div>
+        <div className="row">
+          <div className="input-field col s6">
+            <label htmlFor="city" className="active">
+              Town / City
+            </label>
+            <input
+              onChange={handleChange}
+              value={city}
+              id="city"
+              type="text"
+              className="validate"
+            />
+          </div>
+        </div>
+        <div className="row">
+          <div className="input-field col s6">
+            <select onSubmit={handleSubmit} id="stateAdd">
+              <option value="" disabled selected>
+                Choose your state
+              </option>
+              <option value="Alabama">Alabama</option>
+              <option value="Alaska">Alaska</option>
+              <option value="Arizona">Arizona</option>
+              <option value="Arkansas">Arkansas</option>
+              <option value="California">California</option>
+              <option value="Colorado">Colorado</option>
+              <option value="Connecticut">Connecticut</option>
+              <option value="Delaware">Delaware</option>
+              <option value="Florida">Florida</option>
+              <option value="Georgia">Georgia</option>
+              <option value="Hawaii">Hawaii</option>
+              <option value="Idaho">Idaho</option>
+              <option value="Illinois">Illinois</option>
+              <option value="Indiana">Indiana</option>
+              <option value="Iowa">Iowa</option>
+              <option value="Kansas">Kansas</option>
+              <option value="Kentucky">Kentucky</option>
+              <option value="Louisiana">Louisiana</option>
+              <option value="Maine">Maine</option>
+              <option value="Maryland">Maryland</option>
+              <option value="Massachusetts">Massachusetts</option>
+              <option value="Michigan">Michigan</option>
+              <option value="Minnesota">Minnesota</option>
+              <option value="Mississippi">Mississippi</option>
+              <option value="Missouri">Missouri</option>
+              <option value="Montana">Montana</option>
+              <option value="Nebraska">Nebraska</option>
+              <option value="Nevada">Nevada</option>
+              <option value="New Hampshire">New Hampshire</option>
+              <option value="New Jersey">New Jersey</option>
+              <option value="New Mexico">New Mexico</option>
+              <option value="New York">New York</option>
+              <option value="North Carolina">North Carolina</option>
+              <option value="North Dakota">North Dakota</option>
+              <option value="Ohio">Ohio</option>
+              <option value="Oklahoma">Oklahoma</option>
+              <option value="Oregon">Oregon</option>
+              <option value="Pennsylvania">Pennsylvania</option>
+              <option value="Rhode Island">Rhode Island</option>
+              <option value="South Carolina">South Carolina</option>
+              <option value="South Dakota">South Dakota</option>
+              <option value="Tennessee">Tennessee</option>
+              <option value="Texas">Texas</option>
+              <option value="Utah">Utah</option>
+              <option value="Vermont">Vermont</option>
+              <option value="Virginia">Virginia</option>
+              <option value="Washington">Washington</option>
+              <option value="West Virginia">West Virginia</option>
+              <option value="Wisconsin">Wisconsin</option>
+              <option value="Wyoming">Wyoming</option>
+            </select>
+            <label htmlFor="stateAdd" className="active">
+              State
+            </label>
+          </div>
+          <div className="input-field col s6">
+            <label htmlFor="zip" className="active">
+              Zip Code
+            </label>
+            <input
+              onChange={handleChange}
+              value={zip}
+              id="zip"
+              type="text"
+              className="validate"
+            />
+          </div>
+        </div>
+        <div className="row">
+          <a href="delete the user" className="col s6 delete">
+            Delete this user
+          </a>
+          <button type="button" className="btn waves-effect waves-light col s4">
+            Submit Changes
+          </button>
+        </div>
+      </form>
+    </div>
+  )
+}
+
+const mapStateToProps = state => {
+  return {
+    firstName: state.user.firstName,
+    lastName: state.user.lastName,
+    email: state.user.email,
+    address: state.user.address
+  }
+}
+
+export default connect(mapStateToProps)(UserInfo)
