+++ public/style.css
@@ -183,6 +183,18 @@ User CSS
   width: 400px;
 }
 
+.infoTag {
+  height: 3rem;
+  font-size: 18px;
+  margin: 0 0 8px 0;
+  padding: 0;
+  padding-top: 10px;
+}
+
+.delete {
+  padding-top: 10px;
+}
+
 .card {
   overflow: visible;
 }
